pyproject.toml
[project]
name = "deforestation-early-warning"
version = "1.0.0"
description = "Predict deforestation 90 days in advance using AlphaEarth"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    # Core dependencies
    "earthengine-api>=0.1.400",
    "numpy>=1.24.3,<2.0",
    "pandas>=2.0.3",
    "scipy>=1.11.0",

    # Machine Learning
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.3",
    "lightgbm>=4.1.0",
    "shap>=0.43.0",

    # Time series and feature engineering
    "statsmodels>=0.14.0",
    "ruptures>=1.1.9",

    # Geospatial
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "folium>=0.15.0",
    "geopy>=2.4.0",

    # Visualization
    "matplotlib>=3.7.1",
    "seaborn>=0.12.2",
    "plotly>=5.17.0",

    # Web framework
    "streamlit>=1.29.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "streamlit-folium>=0.15.0",

    # Utilities
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.1",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "black>=23.10.0",
    "ruff>=0.1.0",
]

stretch = [
    # "modAL-python>=0.4.1",  # Active learning
    # "econml>=0.14.1",       # Causal inference
    # "ortools>=9.7.2996",    # Optimization
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
